+++ packages/toolpad-core/src/createComponent.tsx
@@ -1,21 +1,5 @@
-import * as ReactIs from 'react-is';
-import { hasOwnProperty } from '@mui/toolpad-utils/collections';
 import { TOOLPAD_COMPONENT } from './constants.js';
 import { ArgTypeDefinition, ComponentConfig, PropValueType, ToolpadComponent } from './types.js';
-
-export function isToolpadComponent(
-  maybeComponent: unknown,
-): maybeComponent is ToolpadComponent<any> {
-  if (
-    !ReactIs.isValidElementType(maybeComponent) ||
-    typeof maybeComponent === 'string' ||
-    !hasOwnProperty(maybeComponent, TOOLPAD_COMPONENT)
-  ) {
-    return false;
-  }
-
-  return true;
-}
 
 type MaybeLegacyArgTypeDefinition = ArgTypeDefinition & { typeDef?: PropValueType };
 
@@ -45,15 +29,5 @@ export default function createComponent<P extends object>(
   }
   return Object.assign(Component, {
     [TOOLPAD_COMPONENT]: config || { argTypes: {} },
-  });
-}
-
-export function getArgTypeDefaultValue<V>(argType: ArgTypeDefinition<{}, V>): V | undefined {
-  return argType.default ?? argType.defaultValue ?? undefined;
-}
-
-export function createToolpadComponentThatThrows(error: Error) {
-  return createComponent(() => {
-    throw error;
   });
 }
