+++ packages/toolpad-core/src/index.tsx
@@ -5,7 +5,18 @@ export type {
   NodeErrorProps,
   Components,
 } from './runtime.js';
-export { Placeholder, Slots, useNode, ComponentsContext } from './runtime.js';
+export {
+  Placeholder,
+  Slots,
+  useNode,
+  ComponentsContext,
+  isToolpadComponent,
+  getArgTypeDefaultValue,
+  createToolpadComponentThatThrows,
+  useComponents,
+  ComponentsContextProvider,
+  useComponent,
+} from './runtime.js';
 
 export type FlowDirection = 'row' | 'column' | 'row-reverse' | 'column-reverse';
 
@@ -17,8 +28,6 @@ export { default as createComponent } from './createComponent.js';
 export * from './createComponent.js';
 
 export * from './types.js';
-
-export * from './componentsContext.js';
 
 export { default as createQuery } from './createQuery.js';
 export * from './createQuery.js';
