+++ packages/toolpad-core/src/server.ts
@@ -1,4 +1,3 @@
-export * from './constants.js';
 export { default as createQuery } from './createQuery.js';
 export * from './createQuery.js';
